requirements.txt
discord.py
matplotlib
Pillow
firebase-admin
